changed identity columns to Long; modified entity manager factory, using JPA instead of hibernate

+++ src/main/db/02_coin_sentiment_setup.sql
@@ -3,14 +3,14 @@ use crypto;
 drop table if exists CoinSentiment cascade;
 
 create table CoinSentiment (
-	Id int not null auto_increment,
+	Id bigint not null auto_increment,
     Symbol varchar(50) not null,
     TwitterChange_24h decimal(10,2),
-    TwitterVolume_24h long,
+    TwitterVolume_24h bigint,
     RedditChange_24h decimal(10,2),
-    RedditVolume_24h long,
+    RedditVolume_24h bigint,
     SocialVolumeChange_24h decimal(10,2),
-    SocialVolume_24h long,
+    SocialVolume_24h bigint,
     Timestamp datetime,
     BatchNum int not null,
     primary key (Id)
